requirements.txt
airefinery-sdk==1.18.1
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.10.0